# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory API Documentation",
        "description": "API documentation for the Inventory project",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Inventory API Server"
        }
    ],
    "paths": {
...
        "/api/fournisseurs": {
            "get": {
                "tags": [
                    "Fournisseurs"
                ],
                "summary": "Lister tous les fournisseurs",
                "description": "Récupère la liste des fournisseurs de l'entreprise connectée.",
                "operationId": "f1e6d5720017e58d1f85ade3e7869b14",
                "responses": {
                    "200": {
                        "description": "Succès - Liste récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Fournisseur"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fournisseurs"
                ],
                "summary": "Créer un nouveau fournisseur",
                "operationId": "0ff9f333fae2b0f26ae48a5f460a1ae6",
                "requestBody": {
                    "description": "Données du fournisseur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Tech Supplier SARL"
                                    },
                                    "contact_nom": {
                                        "type": "string",
                                        "example": "Jean Dupont",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@techsupplier.com",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+237699000000",
                                        "nullable": true
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Rue de la République",
                                        "nullable": true
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Douala",
                                        "nullable": true
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "Cameroun",
                                        "nullable": true
                                    },
                                    "actif": {
                                        "description": "Statut du fournisseur",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Fournisseur principal pour le matériel IT",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès - Fournisseur créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Fournisseur"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fournisseurs/{id}": {
            "get": {
                "tags": [
                    "Fournisseurs"
                ],
                "summary": "Afficher les détails d'un fournisseur",
                "operationId": "5cf212014b1aa572368a1dc946c6fbc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du fournisseur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - Détails du fournisseur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Fournisseur"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Fournisseur introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Fournisseurs"
                ],
                "summary": "Mettre à jour un fournisseur",
                "operationId": "2d19ad51759d6c2e64485e05b98cad74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du fournisseur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Champs à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "contact_nom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "pays": {
                                        "type": "string"
                                    },
                                    "actif": {
                                        "type": "boolean"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès - Fournisseur mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Fournisseur"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Fournisseur introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fournisseurs"
                ],
                "summary": "Supprimer un fournisseur",
                "operationId": "5bc0895daf02ac1ca6531966bdc9688f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du fournisseur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Succès - Fournisseur supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Fournisseur introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Entreprise": {
                "title": "Entreprise",
                "description": "Informations sur une entreprise",
                "properties": {
                    "id_entreprise": {
                        "description": "Identifiant unique de l'entreprise",
                        "type": "string",
                        "format": "uuid"
                    },
                    "company_name": {
                        "description": "Nom de l'entreprise",
                        "type": "string",
                        "maxLength": 255
                    },
                    "tax_number": {
                        "description": "Numéro fiscal",
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "description": "Adresse de l'entreprise",
                        "type": "string",
                        "nullable": true
                    },
                    "company_email": {
                        "description": "Email de l'entreprise",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "logo": {
                        "description": "Logo (URL ou chemin)",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_name": {
                        "description": "Nom du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_email": {
                        "description": "Email du responsable",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "manager_phone": {
                        "description": "Téléphone du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_nui": {
                        "description": "NUI du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Fournisseur": {
                "title": "Fournisseur",
                "description": "Fournisseur de produits",
                "properties": {
                    "id_fournisseur": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_entreprise": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_users": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "contact_nom": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "telephone": {
                        "type": "string",
                        "nullable": true
                    },
                    "adresse": {
                        "type": "string",
                        "nullable": true
                    },
                    "ville": {
                        "type": "string",
                        "nullable": true
                    },
                    "pays": {
                        "type": "string",
                        "nullable": true
                    },
                    "actif": {
                        "type": "boolean"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "entreprise": {
                        "$ref": "#/components/schemas/Entreprise"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Profil": {
                "title": "Profil",
                "properties": {
                    "id_profil": {
                        "description": "Profile's unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nom": {
                        "description": "Profile's name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Profile's description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Utilisateur du système",
                "properties": {
                    "id_users": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_entreprise": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "agence": {
                        "type": "string",
                        "nullable": true
                    },
                    "poste": {
                        "type": "string",
                        "nullable": true
                    },
                    "link_img": {
                        "type": "string",
                        "nullable": true
                    },
                    "matricule": {
                        "type": "string",
                        "nullable": true
                    },
                    "signature": {
                        "type": "string",
                        "nullable": true
                    },
                    "entreprise": {
                        "$ref": "#/components/schemas/Entreprise"
                    },
                    "profil": {
                        "$ref": "#/components/schemas/Profil"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}